+++ 13_TabbarController/LamLT/Bai1/Bai1/Controller/BaseController/BaseTabBarController.swift
@@ -4,6 +4,5 @@ class BaseTabBarController: UITabBarController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
     }
 }
