+++ 13_TabbarController/LamLT/Bai1/Bai1/Controller/ForGotPassWordViewController.swift
@@ -1,30 +1,8 @@
-//
-//  ForGotPassWordViewController.swift
-//  Bai1
-//
-//  Created by PCI0002 on 4/13/20.
-//  Copyright © 2020 LeThanhLam. All rights reserved.
-//
-
 import UIKit
 
-class ForGotPassWordViewController: UIViewController {
+final class ForGotPassWordViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
     }
-
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
 }
