+++ 13_TabbarController/LamLT/Bai1/Bai1/Controller/HomeController/HomeViewController.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-class HomeViewController: UIViewController {
+final class HomeViewController: UIViewController {
 
     @IBOutlet weak var userNameLabel: UILabel!
