+++ 13_TabbarController/LamLT/Bai1/Bai1/Controller/MapController/MapViewController.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-class MapViewController: UIViewController {
+final class MapViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
