+++ 13_TabbarController/LamLT/Bai1/Bai1/Controller/ProfileController/ProfileViewController.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-class ProfileViewController: UIViewController {
+final class ProfileViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
