+++ 14_MVVM/LamLT/BaiTapMVVM/BaiTapMVVM/Controllers/DetailVC/DetailViewModel.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-class DetailViewModel {
+final class DetailViewModel {
     var dataDetail: [DataDetail] = []
     let images = ["image1", "image2", "image3", "image4", "image5"]
