+++ 14_MVVM/LamLT/BaiTapMVVM/BaiTapMVVM/Controllers/ProfileVC/ProfileViewController.swift
@@ -4,9 +4,9 @@ class ProfileViewController: BaseViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
     }
-    
+
     override func setupUI() {
-           super.setupUI()
-           title = "Profile"
-       }
+        super.setupUI()
+        title = "Profile"
+    }
 }
