+++ 14_MVVM/LamLT/BaiTapMVVM/BaiTapMVVM/Models/DataManager.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-class DataManager {
+final class DataManager {
     var name: String
     var image: String
     var address: String
@@ -18,16 +18,4 @@ class DataManager {
     }
 }
 
-final class DataDetail {
-    var name: String
-    var image: String
-    var comment: String
-    var time: String
-    
-    init(name: String = "", image: String = "", comment: String = "", time: String = "") {
-        self.name = name
-        self.image = image
-        self.comment = comment
-        self.time = time
-    }
-}
+
